### 2024-01-13 #6229 @editorialbot
[PRE REVIEW]: CalibrateEmulateSample.jl: Accelerated Parametric Uncertainty Quantification
**Submitting author:** @odunbar (<a href="http://orcid.org/0000-0001-7374-0382">Oliver Dunbar</a>)
**Repository:** https://github.com/CliMA/CalibrateEmulateSample.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.5.0
