### 2024-01-25 2dc501f
add test for MV case, and comment redundant if-else branches
