+++ .github/workflows/Tests.yml
@@ -25,7 +25,7 @@ jobs:
     - name: Set up Julia
       uses: julia-actions/setup-julia@v1
       with:
-        version: 1
+        version: 1.9.3
 
     - name: Install Julia Project Packages
       # we add this ENV varaible to force PyCall to download and use Conda rather than
