+++ .github/workflows/Tests.yml
@@ -80,7 +80,7 @@ jobs:
     - name: Set up Julia
       uses: julia-actions/setup-julia@v1
       with:
-        version: 1.8.5
+        version: 1.9.2
         
     - name: Install Julia Project Packages
       env:
