+++ .github/workflows/Tests.yml
@@ -80,7 +80,7 @@ jobs:
     - name: Set up Julia
       uses: julia-actions/setup-julia@v1
       with:
-        version: 1.9.1
+        version: 1.9.0
                  # later versions causes hanging in package building
     - name: Install Julia Project Packages
       env:
