### 2024-01-29 b6a80c9
Bugfix parameter name bug for MV distributions, cap mac-os tests to julia v1.9.0 (#277)
* sorted parameter name multiplicity for MV distributions

* add test for MV case, and comment redundant if-else branches

* sqrt bug

* missed codecov

* check Mac test with Julia 1.9.3

* wrong version

* try 1.8.5 macos

* try 1.9.2 macos

* try 1.9.0

* 1.9.1 try

* latest working version 1.9.0`
