### 2024-01-30 #6229 @odunbar
Here is a referee list:
ziyiyin97
odow
matt-graham
Shreyas911
thomaspinder

Note that ziyiyin97 & odow also reviewed our sister-package for this codebase (EnsembleKalmanProcesses.jl).
