### 2024-01-31 42ca861
add diags and plotting for it in examples, VERY HACKY
updates from hpc

rm prints

repeats for uq_for_edmf

add save jld2 and lines+series depending on repeats

add jld2 and log-scale

rm prints, typos

format
