### 2024-02-13 3ee2ece
Orad/fix emulator bug repeat call (#281)
* bugfix repeated calls, by skipping build

* typo in test

plots, repeats, more tweaks

format

varying dimension is easier

ens propto dim

adds recompute_cov and inflation/localization for scalar RF

different plots and configs with localization etc.

exp to fit paper results

add statsbase

add plotting when no repeats

add coeffs, hardcode some restart options in scalar RF

add coeffs, hardcode some restart options in scalar RF

kron add regularization matrix

save states and replotting from file

prior plots for l63

added save+plot for GFunc

add
[…]
