+++ examples/EDMF_data/plot_posterior.jl
@@ -9,23 +9,34 @@ using Dates
 # CES 
 using CalibrateEmulateSample.ParameterDistributions
 
+#####
+# Creates 1 plots: One for a specific case, One with 2 cases, and One with all cases (final case being the prior).
+
 
 # date = Date(year,month,day)
 
 # 2-parameter calibration exp
 #exp_name = "ent-det-calibration"
-#date_of_run = Date(2023, 10, 17)
+#date_of_run = Date(2023, 10, 5)
 
 # 5-parameter calibration exp
 exp_name = "ent-det-tked-tkee-stab-calibration"
-date_of_run = Date(2024, 2, 2)
+date_of_run = Date(2024, 06, 14)
 
 # Output figure read/write directory
 figure_save_directory = joinpath(@__DIR__, "output", exp_name, string(date_of_run))
 data_save_directory = joinpath(@__DIR__, "output", exp_name, string(date_of_run))
 
+#case:
+cases = [
+    "GP", # diagonalize, train scalar GP, assume diag inputs
+    "RF-prior",
+    "RF-vector-svd-nonsep",
+]
+case_rf = cases[3]
+
 # load
-posterior_filepath = joinpath(data_save_directory, "posterior.jld2")
+posterior_filepath = joinpath(data_save_directory, "$(case_rf)_posterior.jld2")
 if !isfile(posterior_filepath)
     throw(ArgumentError(posterior_filepath * " not found. Please check experiment name and date"))
 else
@@ -39,15 +50,106 @@ transformed_posterior_samples =
 
 # histograms
 nparam_plots = sum(get_dimensions(posterior)) - 1
-density_filepath = joinpath(figure_save_directory, "posterior_dist_comp.png")
-transformed_density_filepath = joinpath(figure_save_directory, "posterior_dist_phys.png")
+density_filepath = joinpath(figure_save_directory, "$(case_rf)_posterior_dist_comp.png")
+transformed_density_filepath = joinpath(figure_save_directory, "$(case_rf)_posterior_dist_phys.png")
 labels = get_name(posterior)
 
-data = (; [(Symbol(labels[i]), posterior_samples[i, :]) for i in 1:length(labels)]...)
-transformed_data = (; [(Symbol(labels[i]), transformed_posterior_samples[i, :]) for i in 1:length(labels)]...)
+burnin = 50_000
 
-p = pairplot(data => (PairPlots.Scatter(),))
-trans_p = pairplot(transformed_data => (PairPlots.Scatter(),))
+data_rf = (; [(Symbol(labels[i]), posterior_samples[i, burnin:end]) for i in 1:length(labels)]...)
+transformed_data_rf =
+    (; [(Symbol(labels[i]), transformed_posterior_samples[i, burnin:end]) for i in 1:length(labels)]...)
+
+p = pairplot(data_rf => (PairPlots.Contourf(sigmas = 1:1:3),))
+trans_p = pairplot(transformed_data_rf => (PairPlots.Contourf(sigmas = 1:1:3),))
+
+save(density_filepath, p)
+save(transformed_density_filepath, trans_p)
+
+#
+#
+#
+
+case_gp = cases[1]
+# load
+posterior_filepath = joinpath(data_save_directory, "$(case_gp)_posterior.jld2")
+if !isfile(posterior_filepath)
+    throw(ArgumentError(posterior_filepath * " not found. Please check experiment name and date"))
+else
+    println("Loading posterior distribution from: " * posterior_filepath)
+    posterior = load(posterior_filepath)["posterior"]
+end
+# get samples explicitly (may be easier to work with)
[…]
