+++ src/ScalarRandomFeature.jl
@@ -163,6 +163,8 @@ function ScalarRandomFeatureInterface(
         "multithread" => "ensemble", # instead of "tullio"
         "verbose" => false, # verbose optimizer statements
         "accelerator" => EKP.DefaultAccelerator(), # acceleration with momentum
+        "recompute_cov_at" => Inf, # whether to recompute the output_covariance
+        "localization" => EKP.Localizers.NoLocalization(), # localization / sample error correction for small ensembles
     )
 
     if !isnothing(optimizer_options)
@@ -201,6 +203,7 @@ function hyperparameter_distribution_from_flat(
 
     M = zeros(input_dim) #scalar output
     U = hyperparameters_from_flat(x, input_dim, kernel_structure)
+
     if !isposdef(U)
         println("U not posdef - correcting")
         U = posdef_correct(U)
@@ -327,7 +330,6 @@ function build_models!(
     optimizer_options = get_optimizer_options(srfi)
     optimizer = get_optimizer(srfi) # empty vector
 
-
     # Optimize features with EKP for each output dim
     # [1.] Split data into test/train 80/20 
     train_fraction = optimizer_options["train_fraction"]
@@ -347,8 +349,6 @@ function build_models!(
     n_iteration = optimizer_options["n_iteration"]
     diagnostics = zeros(n_iteration, n_rfms)
     for i in 1:n_rfms
-
-
         io_pairs_opt = PairedDataContainer(
             input_values[:, idx_shuffle],
             reshape(output_values[i, idx_shuffle], 1, size(output_values, 2)),
@@ -362,7 +362,7 @@ function build_models!(
         else
             throw(
                 ArgumentError(
-                    "Unknown optimizer option for multithreading, please choose from \"tullio\" (allows Tullio.jl to control threading in RandomFeatures.jl, or \"loops\" (threading optimization loops)",
+                    "Unknown optimizer option for multithreading, please choose from \"tullio\" (allows Tullio.jl to control threading in RandomFeatures.jl), or \"ensemble\" (threading is done over the ensemble)",
                 ),
             )
         end
@@ -409,28 +409,72 @@ function build_models!(
         opt_verbose_flag = optimizer_options["verbose"]
         scheduler = optimizer_options["scheduler"]
         accelerator = optimizer_options["accelerator"]
+        recompute_cov_at = optimizer_options["recompute_cov_at"]
+        localization = optimizer_options["localization"]
 
         initial_params = construct_initial_ensemble(rng, prior, n_ensemble)
-        min_complexity = n_features_opt * log(regularization.λ)
-        min_complexity = sqrt(abs(min_complexity))
-        data = vcat(get_outputs(io_pairs_opt)[(n_train + 1):end], 0.0, min_complexity)
-        ekiobj = EKP.EnsembleKalmanProcess(
-            initial_params,
-            data,
-            Γ,
-            Inversion(),
-            scheduler = scheduler,
-            rng = rng,
-            accelerator = accelerator,
-            verbose = opt_verbose_flag,
-        )
+        data = vcat(get_outputs(io_pairs_opt)[(n_train + 1):end], 0.0, 0.0)
+        ekiobj = [
+            EKP.EnsembleKalmanProcess(
+                initial_params,
+                data,
+                Γ,
+                Inversion(),
+                scheduler = scheduler,
+                rng = rng,
+                accelerator = accelerator,
+                verbose = opt_verbose_flag,
+                localization_method = localization,
+            ),
+        ]
         err = zeros(n_iteration)
 
         # [4.] optimize with EKP
-        for it in 1:n_iteration
+        Γ_old = [Γ]
+        for i in 1:n_iteration
 
             #get parameters:
-            lvec = transform_unconstrained_to_constrained(prior, get_u_final(ekiobj))
[…]
