### 2024-02-13 6fcab82
Orad/fix emulator bug repeat call (#281)
* bugfix repeated calls, by skipping build

* typo in test
