### 2024-02-19 #6372 @editorialbot
[REVIEW]: CalibrateEmulateSample.jl: Accelerated Parametric Uncertainty Quantification
**Submitting author:** @odunbar (<a href="http://orcid.org/0000-0001-7374-0382">Oliver Dunbar</a>)
**Repository:** https://github.com/CliMA/CalibrateEmulateSample.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.5.1
