### 2024-02-27 CliMA/CalibrateEmulateSample.jl#283 @matt-graham
Sinusoid example not fully reproducible
Raising as part of JOSS review openjournals/joss-reviews/issues/6372

The sinusoid example in the documentation https://clima.github.io/CalibrateEmulateSample.jl/dev/examples/sinusoid_example/ (and partially included in the JOSS paper), is not fully reproducible. I think this is at least in part due to the use of unseeded random number generators when sampling the observation noise for the simulated observations in
https://github.com/CliMA/CalibrateEmulateSample.jl/blob/6fcab82a690d71ef7c6b3810901da80ba1cfcb43/docs/src/examples/sinusoid_example.md?plain=1#L97-L98
and when simulating the Markov chain in
https://github.com/CliMA/CalibrateEmulateSample.jl/blob/6fcab82a690d71ef7c6b3810901da80ba1cfcb43/docs/src/examples/sinusoid_example.md?plain=1#L359-L365
so that we get different simulated observations from the values mentioned in the tutorial and different Markov chain samples.

There are also a few other small issues which mean the tutorial isn't runnable as is / a complete self-contained example

- [x] Missing imports - for example `Statistics` for the use of `mean` and `Distributions` for the use of `MvNormal`.
- [x] `emulator` variable passed to `MCMCWrapper` not defined (presumably one of `emulator_gp` or `emulator_random_features` should be aliased to this?)
- [x] `init_sample` variable passed as `init_params` argument to `MCMCWrapper` is not defined
- [x] In the random features example there are lots of undefined variables - `nugget`, `n_features`, `input_dim`, `output_dim` (presumably the last two might map to `dim_params` and `dim_output`)
