### 2024-02-27 CliMA/CalibrateEmulateSample.jl#284 @matt-graham
Author list on JOSS paper
Raising as part of JOSS review openjournals/joss-reviews/issues/6372

One of the points on the reviewed checklist is

> Does the full list of paper authors seem appropriate and complete?

and the [JOSS ethics guidelines ](https://joss.theoj.org/about#ethics) state that

> Author lists must be correct and complete. All listed authors must have made a contribution to the work, and all significant contributors should be included in the author list.

As there are a few GitHub users who appear from the [GitHub contributors data](https://github.com/CliMA/CalibrateEmulateSample.jl/graphs/contributors) to have made relatively significant contributions to the software who do not appear to be included as authors (namely `dburov190`, `charleskawczynski` and `tsj5` though apologies if I am mismatching names to GitHub usernames) compared to other contributors who have been included as authors, can I check if there has been a discussion with those contributors about authorship and/or if there is some reason for not including them on the author list?
