### 2024-02-27 CliMA/CalibrateEmulateSample.jl#285 @matt-graham
Role of random phase in sinusoid example unclear
Raising as part of JOSS review openjournals/joss-reviews/issues/6372

In the model definition in the sinusoid example included in the documentation and JOSS paper

https://github.com/CliMA/CalibrateEmulateSample.jl/blob/6fcab82a690d71ef7c6b3810901da80ba1cfcb43/docs/src/examples/sinusoid_example.md?plain=1#L66-L69

https://github.com/CliMA/CalibrateEmulateSample.jl/blob/6fcab82a690d71ef7c6b3810901da80ba1cfcb43/docs/src/examples/sinusoid_example.md?plain=1#L110-L114

it is not clear to me what the role of the random phase $\phi$ is and whether the model as defined fits into the framework assumed in the original _Calibrate, Emulate, Sample_ (CES) paper and described in the JOSS paper and documentation.

The CES paper defines the generative model for the observations $y \in \mathbb{R}^d$ given parameters $\theta \in \mathbb{R}^p$ as

$$ y = \mathcal{G}(\theta) + \eta, \quad  \eta \sim N(0, \Gamma_y) $$

with the forward map $\mathcal{G} : \mathbb{R}^p \to \mathbb{R}^d$ appearing to be assumed throughout to be deterministic.

The `G` / $\mathcal{G}$ function used in the example is in contrast stochastic due to the phase `phi` being randomly generated (using a global random number generator) within the function itself. Is there a reason for using a model which departs from the assumptions of the original paper as the main tutorial example?

I can see the value in being able to calibrate stochastic simulator models which internally make draws from a random number generator as this is typically a harder class of models to work with (and also very prevalent in practice), but its not clear to me if the CES framework straightforwardly extends to this case.

Relatedly there is a reference to approximate Bayesian computation (ABC) methods in the JOSS paper

> In ABC, this approximation comes from bypassing the likelihood that is usually required in sampling methods, such as MCMC. Instead, the goal ABC is to replace the likelihood with a scalar-valued sampling objective that compares model and data. In CES, the approximation comes from learning the parameter-to-data map, then following this it calculates an explicit likelihood and uses exact sampling via MCMC

In the context of the above point about the use of a stochastic simulator model in the tutorial example I think the differences between CES and ABC are a bit muddied. ABC is typically applied to generative models for which we can simulate observable outputs given parameter values but cannot evaluate a density function for the joint distribution on observable outputs and parameters. ABC methods use a proxy for the density of the conditional distribution of the observables given parameters (likelihood) based on a distance measure between simulated observable outputs and observed data, typically using summary statistics of the observables rather than computing a distance on the full model outputs directly. Typically the assumption is that we can only compute a stochastic estimate of this proxy by sampling from the generative model.

The (random phase) sinusoid example model seems to fit in this description and so be a good candidate for applying ABC methods to. The joint distribution of $\mathcal{G}(\theta), \theta$ for this example is intractable (or at least non-trivial) to evaluate due to the random phase $\phi$. $\mathcal{G}$ effectively involves computing summary statistics of the full sinusoidal output, with the assumed Gaussian likelihood then equivalent to computing a Euclidean distance between the simulated and observed summary statistics, and as in ABC we only able to evaluate a stochastic estimate of this likelihood.

Given the additional complexity arising from the use of a stochastic simulator model, and it being unclear if / how the CES framework applies to such models, I would personally recommend just omitting the random phase aspect from the sinusoid example (by for example drawing a fixed random value for the phase `phi` _outside_ the `model` function).
