### 2024-03-05 d91eece
* Copied over code that tracks random number generator `rng` that was in src/examples/Sinusoid/, so now the docs page can be run and is reproducible. This also changed a few numbers slightly that were initially run through the examples pages rather than the docs pages. * `rng` is now saved into `.jld2` files at the end of the example scripts and then opened at the start of the next script, so entire workflow agrees with results of running snippets from the docs pages. * I increased the nugget term from 1e-12 to 1e-9 to help stability and consistency. * Added missing imports Statistics and Dist
[…]
