+++ docs/src/examples/sinusoid_example.md
@@ -1,7 +1,10 @@
 # Sinusoid Example
 
+!!! info "How do I run this code?"
+    The full code is found in the [`examples/`](https://github.com/CliMA/CalibrateEmulateSample.jl/tree/main/examples) directory of the github repository
+
 ## Background
-This example demonstrates how to use CalibrateEmulateSample.jl for a simple model that generates noisy
+This example demonstrates how to use `CalibrateEmulateSample.jl` for a simple model that generates noisy
 observables of a signal. The sinusoid signal is defined by two parameters: its shift along the vertical axis
 and its amplitude. We make noisy observations of the signal and we can calculate the mean of the signal,
 which is informative about its shift along the axis, and the range of the signal, which is informative 
@@ -31,7 +34,7 @@ where $\Gamma$ is the observational covariance matrix. We will assume the noise 
 
 # Walkthrough of code
 
-The code is split into four sections:
+You can find the full scripts to reproduce this tutorial in `examples/Sinusoid/`. The code is split into four sections:
 
 1. Model set up in `sinusoid_setup.jl`
 2. Calibrate in `calibrate.jl`
@@ -49,6 +52,8 @@ First, we load the packages we need for setting up the model:
 using LinearAlgebra, Random
 using Plots
 using JLD2
+using Statistics, Distributions
+
 ```
 
 We define a model that generates a sinusoid given parameters $\theta=(A,v)$ 
@@ -59,11 +64,12 @@ The model adds a random phase shift upon evaluation.
 # Seed for pseudo-random number generator.
 rng_seed = 41
 rng = Random.MersenneTwister(rng_seed)
-# Define x-axis
-dt = 0.01
-trange = 0:dt:(2 * pi + dt)
 
-function model(amplitude, vert_shift)
+function model(amplitude, vert_shift; rng = Random.GLOBAL_RNG)
+    # Define x-axis
+    dt = 0.01
+    trange = 0:dt:(2 * pi + dt)
+    # Set phi, the random phase
     phi = 2 * pi * rand(rng)
     return amplitude * sin.(trange .+ phi) .+ vert_shift
 end
@@ -79,7 +85,7 @@ vert_shift_true = 7.0
 theta_true = [amplitude_true, vert_shift_true]
 dim_params = 2
 # Generate the "true" signal for these parameters
-signal_true = model(amplitude_true, vert_shift_true)
+signal_true = model(amplitude_true, vert_shift_true; rng = rng)
 ```
 We will observe properties of the signal that inform us about the amplitude and vertical 
 position. These properties will be the range (the difference between the maximum and the minimum),
@@ -95,7 +101,7 @@ observables. We call this $y_{obs}$. The user can choose the observational covar
 dim_output = 2
 Γ = 0.2 * I
 white_noise = MvNormal(zeros(dim_output), Γ)
-y_obs = [y1_true, y2_true] .+ rand(white_noise)
+y_obs = [y1_true, y2_true] .+ rand(rng, white_noise)
 println("Observations:", y_obs)
 ```
 This gives $y_{obs}=(6.15, 6.42)$.
@@ -107,9 +113,9 @@ It will be helpful for us to define a function $G(\theta)$, which returns these 
 (the range and the mean) of the sinusoid given a parameter vector. 
 
 ```julia
-function G(theta)
+function G(theta; rng = Random.GLOBAL_RNG)
     amplitude, vert_shift = theta
-    sincurve = model(amplitude, vert_shift)
+    sincurve = model(amplitude, vert_shift; rng = rng)
[…]
