### 2024-03-06 CliMA/CalibrateEmulateSample.jl#283 @odunbar
Hi @matt-graham, thanks for raising these issues!

I'd like to ask was if it clear that the example on our docs page is simply documenting code within [`examples/Sinusoid`](https://github.com/CliMA/CalibrateEmulateSample.jl/tree/main/examples/Sinusoid) on the github? I find that our example code here can be run completely in a self-contained manner, and reproduces all plots etc.

For example
- `init_sample` is defined [here](https://github.com/CliMA/CalibrateEmulateSample.jl/blob/6fcab82a690d71ef7c6b3810901da80ba1cfcb43/examples/Sinusoid/sample.jl#L50)
- `emulator` is loaded [here](https://github.com/CliMA/CalibrateEmulateSample.jl/blob/6fcab82a690d71ef7c6b3810901da80ba1cfcb43/examples/Sinusoid/sample.jl#L37-L38)
- likewise the random features items are all defined [here and afterwards](https://github.com/CliMA/CalibrateEmulateSample.jl/blob/6fcab82a690d71ef7c6b3810901da80ba1cfcb43/examples/Sinusoid/emulate.jl#L89)

We do want the docs to be good teaching points, and be instructive, but I would say the walkthrough, though featuring snippets, is **not intended to be used to reproduce the example**, whereas the code in the examples directory is intended to reproduce everything (inc. plots etc).

In light of this I think perhaps we should also add the item
- [ ] Add a clear link from "documentation of example" -> "full code of example" at the top of example documentation page.

What do you think?
