### 2024-03-06 CliMA/CalibrateEmulateSample.jl#284 @odunbar
Thanks for the comment. I would validate our author-list choices by saying that the selection contains developers for the CES package, who provided either offline-insights, code developments, or both, and wish to be associated and assume levels of stewardship of the codebase. Regarding your choices

- tsj5 has left the project a couple of years ago and although I would have liked them to be an author, they did not respond to email requests from me to be one. As authorship comes with some responsibility over the code.
- dburov helped with the original setup in the code but wished to dissociate themselves from this project several years ago for a different field of research
- charleskawczynski/sjbyrne/jakebolewski are/were CliMA software developers and certainly provided some contribution to initiating the codebase, initial commits, and maintenance e.g. putting GH workflows/formatting etc. in line with CliMA package ecosystem. They appear as significant contributors, and though their contributions are valuable, I would say they are not "associated" with this codebase, any more than the other packages they maintain.
