### 2024-03-06 CliMA/CalibrateEmulateSample.jl#285 @odunbar
The short and unsatisfying answer, is that the phase was included to make the example consistent with the related "simple example" in the `EnsembleKalmanProcesses.jl` package.

The original use was just a light comment about if you learn a shift/amplitude in an inverse problem, you may not need to know all details of the "model" `f`. In particular, the `G` is actually deterministic (up-to numerical discretization error), even though `f` is a stochastic process, and the random phase shift has no affect on the inverse problem itself.  However if instead we took an observation of `f` pointwise, e.g. `f(0)` then I agree this would give a more interesting form of stochastic problem, with non-Gaussian noise and there CES would impose Gaussian likelihood constraints to approximate it.

With this in mind, CES should be applicable here, however there is ambiguity in exposition between the roles of `f` and `G`. - If you would still like us to remove the shift ``\phi`` , then we could still do this.
- Alternatively In the sinusoid docs, we can highlight that our G is deterministic, despite f being stochastic and highlight that in setting up inverse problems, that one does not need to account for all model complexity to learn about parameters.

Let us know what you would prefer.
