### 2024-03-08 CliMA/CalibrateEmulateSample.jl#284 @matt-graham
Hi @odunbar - thanks for the clarifications. That all seem reasonable so closing this.
