### 2024-03-11 #6372 @matt-graham
@sappelhoff - I've currently got no spare capacity due to other commitments so will not be able to look at this again till next week. I saw that the authors responded to the issues I previously raised but haven't had a chance to look through their responses to two of the issues yet. I also have some minor feedback on the paper which I haven't yet had a chance to transcribe in to an issue.
