### 2024-03-16 CliMA/CalibrateEmulateSample.jl#283 @nluetts
Hi :wave:

I am replying to this also in the context of the JOSS review https://github.com/openjournals/joss-reviews/issues/6372

When I looked at your Sinusoid example in the docs, it was in fact not immediately clear to me that it was intended to describe the Sinusoid example in your `examples` folder, because I was looking at the docs first and the repository only a bit later, and there seems to be no mention of the `examples` folder in the respective section in the docs.
So I think a link in the docs plus an explicit statement about the `examples` folder would be very helpful indeed.
