### 2024-03-18 CliMA/CalibrateEmulateSample.jl#283 @matt-graham
Apologies for the delay replying - here. I agree with @nluetts that having a link to the full code in the documentation would be useful - I also missed the `examples` folder initially.
