### 2024-03-19 6ba1335
added link to examples/ dir to docs
