### 2024-03-19 c3a71b4
add subdir for emulator examples to avoid ambiguity
