### 2024-03-20 cde70d3
added link to examples/ dir to docs (#290)
* added link to examples/ dir to docs

* add subdir for emulator examples to avoid ambiguity
