+++ docs/src/examples/emulators/ishigami_3d_1d.md
@@ -1,5 +1,8 @@
 # Global Sensitiviy Analysis for an emulated Ishigami function
 
+!!! info "How do I run this code?"
+    The full code is found in the [`examples/Emulator/`](https://github.com/CliMA/CalibrateEmulateSample.jl/tree/main/examples/Emulator) directory of the github repository
+
 In this example, we assess directly the performance of our machine learning emulators. The task is to learn a function for use in a [global sensitivity analysis](https://en.wikipedia.org/wiki/Variance-based_sensitivity_analysis). In particular, we learn the Ishigami function
 ```math
 f(x; a, b) = (1 + bx_3^4)\sin(x_1) + a \sin(x_2), \forall x\in [-\pi,\pi]^3
