### 2024-03-22 87c069c
Sinusoidal example (#287)
* `rng` is now saved into `.jld2` files at the end of the example scripts and then opened at the start of the next script, so entire workflow agrees with results of running snippets from the docs pages, if you set the same `rng` at the start. But removed `rng` from the docs page to keep it looking clean.
* This also changed a few numbers and plots.
* I increased the nugget term from 1e-12 to 1e-9 to help stability and consistency.
* Added missing imports `Statistics` and `Distributions` in docs and in example file.
* `emulator` -> `emulator_gp` in docs page
* defined 
[…]
