+++ docs/src/examples/sinusoid_example.md
@@ -4,7 +4,7 @@
     The full code is found in the [`examples/`](https://github.com/CliMA/CalibrateEmulateSample.jl/tree/main/examples) directory of the github repository
 
 ## Background
-This example demonstrates how to use CalibrateEmulateSample.jl for a simple model that generates noisy
+This example demonstrates how to use `CalibrateEmulateSample.jl` for a simple model that generates noisy
 observables of a signal. The sinusoid signal is defined by two parameters: its shift along the vertical axis
 and its amplitude. We make noisy observations of the signal and we can calculate the mean of the signal,
 which is informative about its shift along the axis, and the range of the signal, which is informative 
@@ -34,7 +34,7 @@ where $\Gamma$ is the observational covariance matrix. We will assume the noise 
 
 # Walkthrough of code
 
-The code is split into four sections:
+You can find the full scripts to reproduce this tutorial in `examples/Sinusoid/`. The code is split into four sections:
 
 1. Model set up in `sinusoid_setup.jl`
 2. Calibrate in `calibrate.jl`
@@ -52,6 +52,8 @@ First, we load the packages we need for setting up the model:
 using LinearAlgebra, Random
 using Plots
 using JLD2
+using Statistics, Distributions
+
 ```
 
 We define a model that generates a sinusoid given parameters $\theta=(A,v)$ 
@@ -59,15 +61,13 @@ We define a model that generates a sinusoid given parameters $\theta=(A,v)$
 The model adds a random phase shift upon evaluation.
 
 ```julia
-# Seed for pseudo-random number generator.
-rng_seed = 41
-rng = Random.MersenneTwister(rng_seed)
 # Define x-axis
 dt = 0.01
 trange = 0:dt:(2 * pi + dt)
 
 function model(amplitude, vert_shift)
-    phi = 2 * pi * rand(rng)
+    # Set phi, the random phase
+    phi = 2 * pi * rand()
     return amplitude * sin.(trange .+ phi) .+ vert_shift
 end
 
@@ -149,7 +149,7 @@ prior_u1 = PD.constrained_gaussian("amplitude", 2, 1, 0, Inf)
 prior_u2 = PD.constrained_gaussian("vert_shift", 0, 5, -Inf, Inf)
 prior = PD.combine_distributions([prior_u1, prior_u2])
 # Plot priors
-p = plot(prior)
+p = plot(prior, fill = :lightgray)
 ```
 ![prior](../assets/sinusoid_prior.png)
 
@@ -158,9 +158,9 @@ We now generate the initial ensemble and set up the EKI.
 N_ensemble = 10
 N_iterations = 5
 
-initial_ensemble = EKP.construct_initial_ensemble(rng, prior, N_ensemble)
+initial_ensemble = EKP.construct_initial_ensemble(prior, N_ensemble)
 
-ensemble_kalman_process = EKP.EnsembleKalmanProcess(initial_ensemble, y_obs, Γ, EKP.Inversion(); rng = rng)
+ensemble_kalman_process = EKP.EnsembleKalmanProcess(initial_ensemble, y_obs, Γ, EKP.Inversion())
 ```
 
 We are now ready to carry out the inversion. At each iteration, we get the
@@ -181,13 +181,13 @@ Finally, we get the ensemble after the last iteration. This provides our estimat
 final_ensemble = EKP.get_ϕ_final(prior, ensemble_kalman_process)
 
 # Check that the ensemble mean is close to the theta_true
-println("Ensemble mean: ", mean(final_ensemble, dims=2))   # [3.08, 6.37]
+println("Ensemble mean: ", mean(final_ensemble, dims=2))   # [3.05, 6.37]
 println("True parameters: ", theta_true)   # [3.0, 7.0]
 ```
 
 | Parameter         | Truth    | EKI mean |
[…]
