### 2024-03-22 CliMA/CalibrateEmulateSample.jl#283 @odunbar
@nluetts @matt-graham

The latest docs [here](https://clima.github.io/CalibrateEmulateSample.jl/dev/examples/sinusoid_example/)

I believe we have cleared up some reproducibility in the example file (PR287) and also a link at the top from the docs page to the examples (PR290) let us know if you want more done here.
