### 2024-03-26 CliMA/CalibrateEmulateSample.jl#283 @nluetts
The info box in the docs is perfect, now this can hardly be overlooked.
