### 2024-03-26 #6372 @nluetts
@sappelhoff The cold season has taken its toll on me a little, but I resumed the review now and have some time during this week, so I try to provide all my feedback by Friday. Apropos feedback, when I have questions and feedback to the JOSS paper draft itself, am I supposed to open an issue over at https://github.com/CliMA/CalibrateEmulateSample.jl or comment in this issue?
