### 2024-03-26 CliMA/CalibrateEmulateSample.jl#293 @matt-graham
Minor suggestions of improvement to JOSS paper 
Raising issue as part of openjournals/joss-reviews/issues/6372

Some minor comments / suggestions for improvement of JOSS paper

> A julia-language

- [x] Usually when referring to the language I think Julia is capitalized so I think it should be a _A Julia language_ here.

> to constrain the free parameter distribution

- [x] I would say _the free parameter distribution_ is a bit vague here and would suggest the _prior parameter distribution_ itself, as what is being described is I think the prior distribution but also 'prior parameter distribution' I think is relatively self-explanatory even for readers unfamiliar with Bayesian inference jargon.

>constrained joint parameter distribution

- [x] I think it is a bit unclear here what is being referred to by the _joint parameter distribution_. The posterior can be seen as restriction of joint distribution on parameters and observed variables given particular values (data) for the observed variables, but it seems like you might just be referring the prior distribution on the parameters here?

> statistical machine-learning emulators

- [x] I would say just _statistical emulators_ is sufficient and this would be consistent with later usage in the paper.

> when compared with traditional algorithms to draw samples from the joint parameter distribution

- [x] Again _joint parameter distribution_ is unclear here - I would specifically state the posterior distribution on parameters if that is what you mean here. Also _traditional algorithms_ is a bit vague here - it would be better to specifically state which algorithms you are referring to (presumably approaches like Markov chain Monte Carlo or sequential Monte Carlo running directly on the target distribution rather than an emulator).

> variants of Random Walk Metropolis [@Sherlock:2010]

- [x] The reference to Sherlock et al. (2010) seemed a little out of place to me here, as random walk Metropolis predates this (review) article. I would potentially either additional cite Metropolis et al. (1953) _Equation of state calculations by fast computing machines_ which proposed the Metropolis algorithm (using a random walk proposal albeit not a Gaussian one) or possibly just omit a citation altogether. But I think this is quite a subjective point so feel free to ignore.

> _Unfortunately this task is intensely computationally expensive, commonly requiring over $10^5$ evaluations of the expensive computer code_

- [x] I don't think this an unreasonable claim, but ideally it could do with some justification with a reference to a specific example or algorithm.

> that classically this has been unavailable

- [x] This doesn't quite read right to me - maybe an additional _for_ is needed at end?

> _Bayesian Markov Chain Monte Carlo_ -

- [x]  _Chain_ does not need to be capitalized and it is unclear _Bayesian_ qualifier needed here - Markov chain Monte Carlo methods can be applied to Bayesian inference problems, but also to sample from any distribution for which we can evaluate an (unnormalized) density function, and the algorithms used for the Bayesian case can typically be used for the more general setting.

> MCMC

- [x] Acronym should be defined on first usage (Markov chain Monte Carlo is used in full previously but not linked to short form).

> State of the field

- [x] You could potentially additionally mention related non-Julia packages here, for example in Python ecosystem PyVBMC (https://github.com/acerbilab/pyvbmc / https://joss.theoj.org/papers/10.21105/joss.05428) and GPry (https://github.com/jonaselgammal/GPry) both have some overlaps with CalibrateEmulateSimulate.jl in terms of being aimed at allowing accelerated Bayesian calibration of expensive to evaluate computer models by exploiting Gaussian process emulators for the model (or rather the logarithm of the target posterior density function in those cases).
