### 2024-03-26 CliMA/CalibrateEmulateSample.jl#283 @matt-graham
@odunbar thanks - #287 and #290 address all my points so closing this.
