### 2024-03-26 CliMA/CalibrateEmulateSample.jl#285 @matt-graham
Sorry for the delay responding here. Thanks for clarification. I'd missed that `f` is almost determinisitic due to the choice of summary statistics here. I would say your second suggested option of adding some additional detail to the documentation would be fine here, as it was mainly that I was confused by the lack of comment on the random phase and what classes of model the CES framework is applicable to, but keeping thing consistent with an example in another package is I think a good reason to keep things as is.
