### 2024-03-27 CliMA/CalibrateEmulateSample.jl#293 @nluetts
Here is some fruther feedback from my side, also in the context of the JOSS review https://github.com/openjournals/joss-reviews/issues/6372

> p. 1 L 20 "To accelerate and smooth this process" (and L36 "smoothed accelerated sampling problem")

It is not clear to me if "smooth this process" is meant as "makes the sampling easier/less time consuming" or literally, i.e. the emulator smooths when going from parameters to observations.

> p. 1 L 36 (listing the sampling methods)

You do not list Gibbs sampling or HMC/NUTS. When familiarizing myself with MCMC (which I have used only a few times so far) I got the impression that Gibbs sampling and HMC/NUTS sampling techniques are kind of the "recommended" sampling methods at the moment. NUTS is also the default in STAN, if I am not mistaken, and Turing.jl also implements NUTS. Are (some of) these samplers covered by "preconditioned Crank-Nicholson" (I was unsure from a very brief look at the cited paper)? Or is there a reason, in the context of CES, to prefer the samplers you have included in your package?

> p. 2 L53 "order of 10^2 evaluations of the code"

I would change "the code" to "the original computer model". This is what it was called in the summary in the beginning of the paper, and it makes it clearer that you get rid of those expensive model runs.

> p. 2 L61 "ensemble Methods"

Methods with capital M.

> p. 2 L68 "the goal ABC is"

"the goal _of_ ABC is"?

> p. 2 L65 and L72 "GpABC" vs "gpABC"

Should this be the same?

> p. 2 L72
> "ABC can be used in more general contexts than CES, but suffers greater
> approximation error and more stringent assumptions, especially in
> multi-dimensional problems."

Can you provide a reference for this statement? Or is this something that is immediately clear to a statistician? Does it not also depend on the chosen emulator in CES?

> Figure 1

Rather a cosmetic problem: the legend says true amplitude = 6.99, but the text in the paper says 7.0 (seems to be an artifact from the signal being made up of finite amount of samples, `examples/Sinusoid/sinusoid_setup.jl:61`).
Ok, the legend is the mean, the text is the vertical shift, so perhaps one could consider it a property of the model (or finite number of "measurements"), but it is a bit confusing.

I think it would make more sense if the observed range (blue double arrow) would be centered with respect to the observed mean (blue dashed line). But your version can be compared to the true values more easily, so feel free to ignore.

> Figure 4, caption:  "... trained on the re-used calibration pairs"

The points (scatter plots) do not seem to be the same as in Figure 3. Should they be the same?

> p. 5 L104 "histogram of the samples from (...?) is displayed in Figure 5."

Seems to be missing some words, e.g. "MCMC sampling".
