### 2024-04-01 CliMA/CalibrateEmulateSample.jl#294 @odunbar
@nluetts Thank you for your detailed combing, I apologize that you encountered so many typos - these are most certainly on me.

If you don't mind, I have added a check-list to your comments to help me mark them off more easily! I'll open a PR for this to address these soon.
