+++ docs/src/installation_instructions.md
@@ -3,27 +3,33 @@
 ### Installing CalibrateEmulateSample.jl
 
 Currently CalibrateEmulateSample (CES) depends on some external python dependencies 
-including `scikit-learn` wrapped by ScikitLearn.jl, which requires a couple extra 
-installation steps:
+including `scipy` (`1.8.1` works) and `scikit-learn` (`1.1.1` works) that wrapped by ScikitLearn.jl.
 
-First clone the project into a new local repository
+If you have dependencies installed already, then the code can be used by simply entering
 
 ```
-> git clone git@github.com:Clima/CalibrateEmulateSample.jl
-> cd CalibrateEmulateSample.jl
+julia --project
+> ]
+> add CalibrateEmulateSample
 ```
 
-Install and build the project dependencies. Given that CES depends on python packages 
-it is easiest to set the project to use its own 
-[Conda](https://docs.conda.io/en/latest/miniconda.html) environment variable
+One may instead clone the project into a new local repository (using SSH or https link from github), to easily access the CES codebase (e.g. to run our example suite) .
+
+If you do not have the dependencies installed, we have found it is easiest to install them via Julia's "Conda.jl",
+```
+julia --project
+> ]
+> add Conda
+> add CalibrateEmulateSample
+```
+Then install the dependencies by having the project use its own [Conda](https://docs.conda.io/en/latest/miniconda.html) environment variable
 (set by exporting the ENV variable `PYTHON=""`).
 
 ```
 > PYTHON="" julia --project -e 'using Pkg; Pkg.instantiate()'
 ```
 
-
-The `scikit-learn` package (along with `scipy`) then has to be installed if using a Julia project-specific Conda environment:
+This call should build Conda and Pycall. The `scikit-learn` package (along with `scipy`) then has to be installed if using a Julia project-specific Conda environment:
 
 ```
 > PYTHON="" julia --project -e 'using Conda; Conda.add("scipy=1.8.1", channel="conda-forge")'
@@ -32,13 +38,7 @@ The `scikit-learn` package (along with `scipy`) then has to be installed if usin
 ```
 
 See the [PyCall.jl documentation](https://github.com/JuliaPy/PyCall.jl#specifying-the-python-version) 
-for more information about how to configure the local Julia / Conda / Python environment. Typically it will require building in the 
-REPL via
-```julia
-> julia --project
-julia> using Pkg
-julia> Pkg.build("PyCall")
-```
+for more information about how to configure the local Julia / Conda / Python environment. 
 
 To test that the package is working:
 
@@ -63,4 +63,4 @@ julia --project=docs/ -e 'using Pkg; Pkg.instantiate()'
 julia --project=docs/ docs/make.jl
 ```
 
-The locally rendered HTML documentation can be viewed at `docs/build/index.html`.
+The locally rendered HTML documentation can be viewed at `docs/build/index.html`. Occasional figures may only be viewable in the online documentation due to the fancy-url package.
