### 2024-04-01 0f65f28
typo in sinusoid
