+++ docs/src/random_feature_emulator.md
@@ -70,7 +70,7 @@ nonseparable_kernel = NonseparableKernel(cov_structure)
 ```
 where the `cov_structure` implies some imposed user structure on the covariance structure. The basic covariance structures are given by 
 ```julia
-1d_cov_structure = OneDimFactor() # the problem dimension is 1
+oned_cov_structure = OneDimFactor() # the problem dimension is 1
 diagonal_structure = DiagonalFactor() # impose diagonal structure (e.g. ARD kernel)
 cholesky_structure = CholeskyFactor() # general positive definite matrix
 lr_perturbation = LowRankFactor(r) # assume structure is a rank-r perturbation from identity
@@ -103,7 +103,7 @@ Dict(
     "train_fraction" => tf,
 )
 ```
-- Decreasing `csm` (default `10.0`) towards `0.0` directly reduces the number of samples to estimate a covariance matrix in the optimizer, by using a shrinkage estimator - the more shrinkage the more approximation (suggestion, keep shrinkage amount below `0.2`).
+- Decreasing `csm` (default `10.0`) towards `0.0` directly reduces the number of samples to estimate a covariance matrix in the optimizer, by using a shrinkage estimator to improve matrix conditioning. Guide: more samples implies less shrinkage for good conditioning and less approximation error. The amount of shrinkage is returned to user as a value between 0 (no shrinkage) and 1 (shrink to diagonal matrix), suggested that users choose `csm` to keep the shrinkage amount below `0.2`.
 - Increasing `tf` towards `1` changes the train-validate split, reducing samples but increasing cost-per-sample and reducing the available validation data (default `0.8`, suggested range `(0.5,0.95)`).
 
 If optimizer convergence stagnates or is too slow, or if it terminates before producing good results, try:
@@ -116,9 +116,9 @@ Dict(
 )
 ```
 We suggest looking at the [`EnsembleKalmanProcesses`](https://github.com/CliMA/EnsembleKalmanProcesses.jl) documentation for more details; but to summarize
-- Reducing optimizer samples `n_e` and iterations `n_i` reduces computation time.
-- If `n_e` becomes less than the number of hyperparameters, the updates will fail and a localizer must be specified in `loc`.
-- If the algorithm terminates at `T=1` and resulting emulators looks unacceptable one can change or add arguments in `sch` e.g. `DataMisfitController("on_terminate"=continue)`
+- Reducing optimizer samples `n_e` and iterations `n_i` reduces computation time but may limit convergence progress, see [here](https://clima.github.io/EnsembleKalmanProcesses.jl/dev/ensemble_kalman_inversion/#Updating-the-Ensemble).
+- If `n_e` becomes less than the number of hyperparameters, the updates may fail and a localizer must be specified in `loc`, see [here](https://clima.github.io/EnsembleKalmanProcesses.jl/dev/localization/). 
+- If the algorithm terminates at `T=1` and resulting emulators looks unacceptable one can change or add arguments in `sch` e.g. `DataMisfitController("on_terminate"=continue)`, see [here](https://clima.github.io/EnsembleKalmanProcesses.jl/dev/learning_rate_scheduler/)
 
 !!! note
     Widely robust defaults here are a work in progress
@@ -179,8 +179,31 @@ build_default_prior(input_dim, output_dim, vector_default_kernel)
 ### Vector, nonseparable: ``\mathbb{R}^{25} \to \mathbb{R}^{50}`` 
 The following represents the most general kernel case.
 
-!!! note "Use low-rank/diagonls representations where possible"
[…]
