### 2024-04-02 267ebf0
typos in updates, and cleared up sample.md language and MCMCChains utilities
