+++ docs/src/sample.md
@@ -10,9 +10,10 @@ Carlo algorithm](https://en.wikipedia.org/wiki/Markov_chain_Monte_Carlo) (MCMC).
 
 ## User interface
 
-We briefly outline an instance of how one sets up and uses MCMC within the CES package. The user first provides a Protocol (i.e. how one wishes to generate proposals)
+We briefly outline an instance of how one sets up and uses MCMC within the CES package. The user first loads the MCMC module, and provides one of the Protocols (i.e. how one wishes to generate sampling proposals)
 
 ```julia
+using CalibrateEmulateSample.MarkovChainMonteCarlo
 protocol = RWMHSampling() # Random-Walk algorithm
 # protocol = pCNMHSampling() # preconditioned-Crank-Nicholson algorithm
 ```
@@ -82,23 +83,21 @@ remember where methods are defined! Below we describe the relevant parts of
 - Further extended for the needs of CES in [Markov chain Monte
   Carlo](https://en.wikipedia.org/wiki/Markov_chain_Monte_Carlo).
 
-## Classes and methods
-
 ### Sampler
 
 A Sampler is AbstractMCMC's term for an implementation of a MCMC sampling algorithm, along with all its configuration
-parameters. All samplers must inherit from `AbstractMCMC.AbstractSampler`. 
+parameters. All samplers are a subtype of AbstractMCMC's `AbstractSampler`. 
 
 Currently CES only implements the Metropolis-Hastings (MH) algorithm. Because it's so straightforward, much of
 AbstractMCMC isn't needed. We implement two variants of MH with two different Samplers: `RWMetropolisHastings` and
-`pCNMetropolisHastings`, both of which inherit from the `AdvancedMH.MHSampler` base class. The public constructor for
+`pCNMetropolisHastings`, both of which are subtypes of `AdvancedMH.MHSampler`. The constructor for
 both Samplers is [`MetropolisHastingsSampler`](@ref); the different Samplers are specified by passing a
 [`MCMCProtocol`](@ref) object to this constructor.
 
-The MH Sampler classes have only one field, `proposal`, which is the distribution used to generate new MH proposals via
-stochastic offsets to the current parameter values. This is done by
+The `MHSampler` has only one field, `proposal`, the distribution used to generate new MH proposals via
+additive stochastic perturbations to the current parameter values. This is done by
 [AdvancedMH.propose()](https://github.com/TuringLang/AdvancedMH.jl/blob/master/src/proposal.jl), which gets called for
-each MCMC `step()` (below). The difference between our two Samplers is in how this proposal is generated:
+each MCMC `step()`. The difference between Samplers comes from how the proposal is generated:
 
 - [`RWMHSampling`](@ref) does vanilla random-walk proposal generation with a constant, user-specified step size (this
   differs from the AdvancedMH implementation, which doesn't provide for a step size.)
@@ -109,7 +108,7 @@ each MCMC `step()` (below). The difference between our two Samplers is in how th
   (2008)](https://www.worldscientific.com/doi/abs/10.1142/S0219493708002378) and [Cotter et. al.
   (2013)](https://projecteuclid.org/journals/statistical-science/volume-28/issue-3/MCMC-Methods-for-Functions--Modifying-Old-Algorithms-to-Make/10.1214/13-STS421.full).
 
-This is the only difference: generated proposals are then either accepted or rejected according to the same MH criterion
+Generated proposals are then either accepted or rejected according to the same MH criterion
 (in `step()`, below.)
 
 ### Models
@@ -121,7 +120,7 @@ likelihood (see [Emulators](@ref)) together with the prior. This is constructed 
 ### Sampling with the MCMC Wrapper object
 
[…]
