+++ docs/src/API/GaussianProcess.md
@@ -8,7 +8,7 @@ CurrentModule = CalibrateEmulateSample.Emulators
 GaussianProcessesPackage
 PredictionType
 GaussianProcess
-GaussianProcess(package::GPPkg)
+GaussianProcess(package::GPPkg)  where {GPPkg <: GaussianProcessesPackage}
 build_models!(::GaussianProcess{GPJL}, ::PairedDataContainer{FT}) where {FT <: AbstractFloat}
 optimize_hyperparameters!(::GaussianProcess{GPJL})
 predict(::GaussianProcess{GPJL},  ::AbstractMatrix{FT}) where {FT <: AbstractFloat}
