### 2024-04-02 cbe8292
API more types
