### 2024-04-02 CliMA/CalibrateEmulateSample.jl#294 @odunbar
I hope the above PR was able to address most of the issues raised.
Thanks again for the time you placed in going through the many pages. I leave here one or two clarifying comments.
- Regarding trade-offs of Emulators: I have only added a note saying that users should give both a go to see what is effective. I don't want to explicitly talk about what is better as these are commonly problem specfic, furthermore the packages do evolve over time and change in performance in different cases.
- Regarding `csm`, it is a number of samples needed to approximate a covariance matrix. The 0.2 is a measure between 0 and 1 of the amount of shrinkage. With huge amounts of samples the shrinkage = 0, with very few samples shrinkage = 1. This is now explained in more detail

The following are not yet resolved issues - as they involve modifying code that is not just in `docs/` and so will be completed in later PRs
- *Requires change/rerun of plots for sinusoid_example* For comparison it would be better if the color scale would cover the same range from figure to figure, especially when plotting the errors.
- *Requires change/rerun of plots for sinusoid_example* Maybe add a table with the numeric results, like in the Gaussian process section. This would help to compare the results from Gaussian process and random features.
- *Requires change in Emulator.jl* The importance of providing the covariance matrix is a bit buried here (not mentioned in the examples, I think). Perhaps display a warning (@warn) if no covariance matrix is provided when constructing the Emulator?
- *Requires change in Emulator.jl* The function Emulator (constructor of type Emulator) does not have an API reference, only the type has an API reference, which is a little unfortunate because one can miss the docs in https://clima.github.io/CalibrateEmulateSample.jl/dev/emulate/#Diagonalization-and-output-dimension-reduction when going directly to the API reference.
