### 2024-04-03 CliMA/CalibrateEmulateSample.jl#293 @odunbar
Hi both - thanks for these comments.
I shall create a PR shortly for to begin addressing them in the markdown. I have added check-boxes to the comments to aid in tracking them.
