### 2024-04-03 cea7662
JOSS docs updates (#296)
* installation instructions updates

* typo in sinusoid

* added better links for julia-free

* contributing emulators

* additional noise cov

* rf improvements

* typos in updates, and cleared up sample.md language and MCMCChains utilities

* typos listed by referee

* try API for GP

* try API GP

* API check again

* API more types

* more detail

* typo
