+++ paper.bib
@@ -272,3 +272,37 @@ urldate = {2024-01-05},
         year = {2023},
         doi = {10.22541/essoar.170365299.96491153/v1},
 }
+
+
+@article{Metropolis:1953,
+  title={Equation of state calculations by fast computing machines},
+  author={Metropolis, Nicholas and Rosenbluth, Arianna W and Rosenbluth, Marshall N and Teller, Augusta H and Teller, Edward},
+  journal={The journal of chemical physics},
+  volume={21},
+  number={6},
+  pages={1087--1092},
+  year={1953},
+  publisher={American Institute of Physics}
+}
+
+@article{Huggins:2023,
+doi = {10.21105/joss.05428},
+url = {https://doi.org/10.21105/joss.05428},
+year = {2023},
+publisher = {The Open Journal},
+volume = {8},
+number = {86},
+pages = {5428},
+author = {Bobby Huggins and Chengkun Li and Marlon Tobaben and Mikko J. Aarnos and Luigi Acerbi},
+title = {PyVBMC: Efficient Bayesian inference in Python},
+journal = {Journal of Open Source Software}
+}
+
+@article{Gammal:2022,
+      title={Fast and robust Bayesian Inference using Gaussian Processes with GPry}, 
+      author={Jonas El Gammal and Nils Schöneberg and Jesús Torrado and Christian Fidler},
+      year={2022},
+      eprint={2211.02045},
+      archivePrefix={arXiv},
+      primaryClass={astro-ph.CO}
+}
