### 2024-04-03 CliMA/CalibrateEmulateSample.jl#293 @odunbar
@nluetts as a couple of responses to your questions.
- by smoothing, we mean regularize the problem (as the statistical emulator mean function will be smoother than the forward map evaluations) i've tried to make this more clear
- RE other methods, we are actually getting a student to look at NUTS/HMC. So it is great that you raise this. The reason we did not initially highlight them is that they are derivative-based methods and so are not directly applicable to "original computer models" that do not have derivatives. However, it should be noted that (precisely due to the smoothing/regularizing property mentioned above) you should be able to regain the ability to use derivative-based methods. If you are interested we are also looking at "Barker proposals" that also are claimed to overcome a lot of the headaches of choosing algorithm parameters for HMC and are remarkably simple to implement compared with NUTS. I shall include a note about the methods.
