+++ paper.bib
@@ -306,3 +306,24 @@ journal = {Journal of Open Source Software}
       archivePrefix={arXiv},
       primaryClass={astro-ph.CO}
 }
+
+@article{livingstone:2022,
+  title={The Barker proposal: combining robustness and efficiency in gradient-based MCMC},
+  author={Livingstone, Samuel and Zanella, Giacomo},
+  journal={Journal of the Royal Statistical Society Series B: Statistical Methodology},
+  volume={84},
+  number={2},
+  pages={496--523},
+  year={2022},
+  publisher={Oxford University Press}
+}
+
+@article{hoffman:2014,
+  title={The No-U-Turn sampler: adaptively setting path lengths in Hamiltonian Monte Carlo.},
+  author={Hoffman, Matthew D and Gelman, Andrew and others},
+  journal={J. Mach. Learn. Res.},
+  volume={15},
+  number={1},
+  pages={1593--1623},
+  year={2014}
+}
