### 2024-04-03 fb4fa50
JOSS paper corrections (#297)
* first comment batch

* addressed further typos

* derivative MCMC

* refs
