### 2024-04-03 340bfdd
updated color limits on the contour plots
