+++ paper.bib
@@ -12,6 +12,14 @@
   journal   = {{SIAM} Review}
 }
 
+@book{Sisson:2018,
+title = {Handbook of {Approximate} {Bayesian} {Computation}},
+isbn = {978-1-351-64346-7},
+publisher = {CRC Press},
+author = {Sisson, Scott A. and Fan, Yanan and Beaumont, Mark},
+year = {2018}
+}
+
 @article{Cleary:2021,
 title = {Calibrate, emulate, sample},
 journal = {Journal of Computational Physics},
