### 2024-04-04 ae4a899
Table for RF to directly compare values to GP
