### 2024-04-04 aa5033c
emulator API
