+++ docs/src/API/Emulators.md
@@ -6,7 +6,7 @@ CurrentModule = CalibrateEmulateSample.Emulators
 
 ```@docs
 Emulator
-Emulator(::MachineLearningTool, ::PairedDataContainer{FT})
+Emulator(::MachineLearningTool, ::PairedDataContainer{FT})  where {FT <: AbstractFloat}
 optimize_hyperparameters!(::Emulator)
 predict
 normalize
