+++ docs/src/API/Emulators.md
@@ -6,7 +6,16 @@ CurrentModule = CalibrateEmulateSample.Emulators
 
 ```@docs
 Emulator
-Emulator(::MachineLearningTool, ::PairedDataContainer{FT})  where {FT <: AbstractFloat}
+function Emulator(
+    ::MachineLearningTool,
+    ::PairedDataContainer{FT};
+    ::Union{AbstractMatrix{FT}, UniformScaling{FT}, Nothing},
+    ::Bool,
+    ::Bool,
+    ::Union{AbstractVector{FT}, Nothing},
+    ::Bool,
+    ::FT,
+) where {FT <: AbstractFloat}
 optimize_hyperparameters!(::Emulator)
 predict
 normalize
