### 2024-04-04 2827e71
JOSS review: sinusoid examples and plots in paper (#298)
* Paper and docs Fig. 1 comparing true and observed signals: changed 6.99 -> 7.0 in legend; edited blue vertical line with arrows for observation range so it is now centered on the observed mean following reviewer comments.

* Paper and docs for contour plots showing how sinusoid outputs vary with respect to parameters, updated color limits to be fixed for all contour plots for easier comparisons

* Docs for the sampling using Random Features: add a table with mean and std of parameter values to directly compare values to GP results.

* 
[…]
