+++ paper.bib
@@ -12,6 +12,25 @@
   journal   = {{SIAM} Review}
 }
 
+@book{Sisson:2018,
+title = {Handbook of {Approximate} {Bayesian} {Computation}},
+isbn = {978-1-351-64346-7},
+publisher = {CRC Press},
+author = {Sisson, Scott A. and Fan, Yanan and Beaumont, Mark},
+year = {2018}
+}
+
+@incollection{Nott:2018,
+  author  = {Nott, David J. and Ong, Victor M.-H. and Fan, Y. and Sisson, S. A.},
+  title = {High-{Dimensional} {ABC}},
+  booktitle   = {Handbook of {Approximate} {Bayesian} {Computation}},
+  isbn = {978-1-315-11719-5},
+  chapter = {8},
+  pages   = {211-241},
+  year    = {2018},
+  publisher = {CRC Press},
+}
+
 @article{Cleary:2021,
 title = {Calibrate, emulate, sample},
 journal = {Journal of Computational Physics},
