### 2024-04-05 5883125
JOSS: Emulator docstring and warning (#299)
* emulator updates

* emulator API

* update test_logs

* test fix

* another test_logs...

* API again...

* rm func

* try

* typo

* need to upd docs in another PR

* try

* final push, docs in next PR
