+++ dev/assets/documenter.js
@@ -4,6 +4,7 @@ requirejs.config({
     'highlight-julia': 'https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.8.0/languages/julia.min',
     'headroom': 'https://cdnjs.cloudflare.com/ajax/libs/headroom/0.12.0/headroom.min',
     'jqueryui': 'https://cdnjs.cloudflare.com/ajax/libs/jqueryui/1.13.2/jquery-ui.min',
+    'minisearch': 'https://cdn.jsdelivr.net/npm/minisearch@6.1.0/dist/umd/index.min',
     'katex-auto-render': 'https://cdnjs.cloudflare.com/ajax/libs/KaTeX/0.16.8/contrib/auto-render.min',
     'jquery': 'https://cdnjs.cloudflare.com/ajax/libs/jquery/3.7.0/jquery.min',
     'headroom-jquery': 'https://cdnjs.cloudflare.com/ajax/libs/headroom/0.12.0/jQuery.headroom.min',
@@ -102,10 +103,9 @@ $(document).on("click", ".docstring header", function () {
   });
 });
 
-$(document).on("click", ".docs-article-toggle-button", function (event) {
+$(document).on("click", ".docs-article-toggle-button", function () {
   let articleToggleTitle = "Expand docstring";
   let navArticleToggleTitle = "Expand all docstrings";
-  let animationSpeed = event.noToggleAnimation ? 0 : 400;
 
   debounce(() => {
     if (isExpanded) {
@@ -116,7 +116,7 @@ $(document).on("click", ".docs-article-toggle-button", function (event) {
 
       isExpanded = false;
 
-      $(".docstring section").slideUp(animationSpeed);
+      $(".docstring section").slideUp();
     } else {
       $(this).removeClass("fa-chevron-down").addClass("fa-chevron-up");
       $(".docstring-article-toggle-button")
@@ -127,7 +127,7 @@ $(document).on("click", ".docs-article-toggle-button", function (event) {
       articleToggleTitle = "Collapse docstring";
       navArticleToggleTitle = "Collapse all docstrings";
 
-      $(".docstring section").slideDown(animationSpeed);
+      $(".docstring section").slideDown();
     }
 
     $(this).prop("title", navArticleToggleTitle);
@@ -224,465 +224,224 @@ $(document).ready(function () {
 
 })
 ////////////////////////////////////////////////////////////////////////////////
-require(['jquery'], function($) {
+require(['jquery', 'minisearch'], function($, minisearch) {
 
-$(document).ready(function () {
-  let meta = $("div[data-docstringscollapsed]").data();
+// In general, most search related things will have "search" as a prefix.
+// To get an in-depth about the thought process you can refer: https://hetarth02.hashnode.dev/series/gsoc
 
-  if (meta?.docstringscollapsed) {
-    $("#documenter-article-toggle-button").trigger({
-      type: "click",
-      noToggleAnimation: true,
-    });
-  }
+let results = [];
+let timer = undefined;
+
+let data = documenterSearchIndex["docs"].map((x, key) => {
+  x["id"] = key; // minisearch requires a unique for each object
+  return x;
 });
 
-})
-////////////////////////////////////////////////////////////////////////////////
-require(['jquery'], function($) {
+// list below is the lunr 2.1.3 list minus the intersect with names(Base)
+// (all, any, get, in, is, only, which) and (do, else, for, let, where, while, with)
+// ideally we'd just filter the original list but it's not available as a variable
+const stopWords = new Set([
+  "a",
+  "able",
+  "about",
+  "across",
+  "after",
[…]
