### 2024-04-05 CliMA/CalibrateEmulateSample.jl#293 @odunbar
I believe we have addressed everything, Please close the issue if you are satisfied.
