### 2024-04-05 #6372 @odunbar
Hi everyone! Thanks for all the comments on the documentation, examples and manuscript.
I believe as of now we have addressed all points raised, with PRs documenting all changes that link to the issues above. Let me know if any item was missed.

I'd like to make one additional change related to adding one further author to the paper. @nefrathenrici .
Though they have not made visible direct contributions to the exact CES repository so far, Nat has been involved in a lot of recent decisionmaking, and added other auxilliary code such as pipelines to enable automated workflows for CalibrateEmulateSample needs for HPC applications, with various PRs to our sister repositories. I think as Nat is also primed to play more of a role moving forwards with CalibrateEmulateSample.jl they will prove a responsible and reliable steward for this codebase in future.

I will link an issue/PR adding Nat to the code here once this is done.

I think our docs and examples have definitely benefited from your reviews, and should hopefully help future users!

Cheers,
Ollie
