### 2024-04-05 CliMA/CalibrateEmulateSample.jl#293 @nluetts
@odunbar thanks for clarifying, and thanks for the pointer to "Barker proposal", which I did not know about so far.

In this new sentence:

> Two such examples, written in python, approximate the log-posterior distribution ...

you could capitalize Python.

@matt-graham the issue can be closed from my side :+1:
