### 2024-04-05 #6372 @nluetts
@sappelhoff : @odunbar addressed all points I raised with regard to the paper and the documentation of the package, so I ticked off the remaining items on my checklist. I agree with @Vaibhavdixit02 conclusions and recommend this paper for acceptance.
