+++ paper.md
@@ -28,6 +28,8 @@ authors:
   - name: Gregory L. Wagner
     orcid: 0000-0001-5317-2445
     affiliation: 5
+  - name: N. Efrat-Henrici
+    affiliation: 1
 
 affiliations:
  - name: Geological and Planetary Sciences, California Institute of Technology
