### 2024-04-05 32623c4
JOSS fix typos (#302)
* typo

* new pdf
