### 2024-04-08 #6372 @sappelhoff
@nluetts I could not find you in our reviewer database. If you enjoyed the process of reviewing for JOSS, we'd be very grateful if you considered signing up for potential future services as a reviewer🙂 ... You can do that here: https://reviewers.joss.theoj.org/join
