### 2024-04-16 #6372 @sappelhoff
@odunbar sorry for the delay. Here is a list of comments from my side. I would appreciate a response to all of them (preferably with links to the commits in which you addressed them), and if you don't want to follow my suggestions on a particular point, I'd appreciate a reason for why not.

- [ ] Are all author ORCIDs added? It's fine if some of the authors don't have an ORCID but if they do, it'd be good to link them.
    - [ ] For example, is this the orcid of one of your co-authors? --> https://orcid.org/0000-0002-9906-7824
- [ ] Please there are [several missing DOIs](https://github.com/openjournals/joss-reviews/issues/6372#issuecomment-2059357366), could you please:
    - [ ] correct the invalid DOIs
    - [ ] go through the list of missing DOIs and for each item make sure that indeed there is no DOI (especially for those items where the bot made suggestions)
- [ ] Please provide consistent names for your co-authors, for example `N. Efrat-Henrici` is the only co-author with an abbreviated first name
- [ ] Please consistently use abbreviations like `e.g.` with or without comma after. Currently, you sometimes use `e.g.,` and sometimes just `e.g.`. I would recommend to follow the APA guidelines on this, which say to use latin abbreviations (also `i.e.`) only in parentheses (**e.g.,** here) and with commas, and not in flowing text like, **for example**, here. See: https://apastyle.apa.org/style-grammar-guidelines/abbreviations/latin ... ultimately this is up to you, but please be consistent
- [ ] you have some occurrences of nested parentheses due to citations like this one: ` (e.g., the Ensemble Kalman65
Sampler (Dunbar, Lopez-Gomez, et al., 2022; Garbuno-Inigo et al., 2020))` please reformat them to not use nested parentheses --> `(e.g., the Ensemble Kalman65
Sampler; Dunbar, Lopez-Gomez, et al., 2022; Garbuno-Inigo et al., 2020)`
- [ ] The labels and ticklabels in your plots are rather small, could you please increase their sizes where possible without changing the plots substantially? Especially for Figure 5
- [ ] I suggest you convert the list of refs in parentheses in the `Research projects using the package` section into a bullet list for improved readability and to avoid the awkward parentheses.
- [ ] In the intro of your paper, the following part would benefit from a three-part enumerated list in my opinion: `the pipeline has three stages, most easily explained in reverse ........`
