### 2024-04-17 #6372 @odunbar
Thank you for the comments, I'll get to it.

Regarding the author name, I asked them to provide a name that they wished to be stated on the manuscript and this is what they provided, which I respected. I'm happy to ask them again (but I am only really comfortable changing it if they are however).
