### 2024-04-18 18d88b8
addressed e.g.s, reseearch list, and CES pipeline list
