### 2024-04-22 c154dc3
replot sinusoid figures with larger fonts
