+++ paper.bib
@@ -16,7 +16,7 @@ title = {Handbook of {Approximate} {Bayesian} {Computation}},
 isbn = {978-1-351-64346-7},
 publisher = {CRC Press},
 author = {Sisson, Scott A. and Fan, Yanan and Beaumont, Mark},
-year = {2018}
+year = {2018},
 doi = {10.1201/9781315117195}
 }
 
@@ -81,7 +81,7 @@ year = {2022}
   volume={2},
   number={3},
   year={2006},
-  publisher={MIT press Cambridge, MA}
+  publisher={MIT press Cambridge, MA},
   doi = {10.1142/S0129065704001899}
 }
 
@@ -93,7 +93,7 @@ year = {2022}
   number={4},
   pages={045001},
   year={2013},
-  publisher={IOP Publishing}
+  publisher={IOP Publishing},
   doi={10.1088/0266-5611/29/4/045001}
 }
 
@@ -114,7 +114,7 @@ year = {2022}
   booktitle={2008 46th Annual Allerton Conference on Communication, Control, and Computing},
   pages={555--561},
   year={2008},
-  organization={IEEE}
+  organization={IEEE},
   doi = {10.1109/allerton.2008.4797607}
 }
 
@@ -151,7 +151,7 @@ doi = {10.1214/13-STS421},
  publisher = {Institute of Mathematical Statistics},
  title = {The Random Walk Metropolis: Linking Theory and Practice Through a Case Study},
  volume = {25},
- year = {2010}
+ year = {2010},
  doi = {10.1214/10-sts327}
 }
 
@@ -296,7 +296,7 @@ journal = {Inverse Problems},
   number={6},
   pages={1087--1092},
   year={1953},
-  publisher={American Institute of Physics}
+  publisher={American Institute of Physics},
   doi = {10.1063/1.1699114}
 }
 
@@ -333,7 +333,7 @@ journal = {Journal of Cosmology and Astroparticle Physics},
   number={2},
   pages={496--523},
   year={2022},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
   doi={10.1111/rssb.12482}
 }
 
