+++ examples/Sinusoid/emulate.jl
@@ -179,15 +179,7 @@ p2 = contour(
     ylabel = "Vertical Shift",
     title = "True Sinusoid Mean",
 )
-p = plot(
-    p1,
-    p2,
-    size = (600, 300),
-    layout = (1, 2),
-    guidefontsize = 14,
-    tickfontsize = 12,
-    legendfontsize = 12,
-)
+p = plot(p1, p2, size = (600, 300), layout = (1, 2), guidefontsize = 14, tickfontsize = 12, legendfontsize = 12)
 savefig(p, joinpath(data_save_directory, "sinusoid_groundtruth_contours.png"))
 # The first panel shows how the range varies with respect to the two parameters in the Gaussian process
 # emulator. The contours show the range is mostly dependent on the amplitude, with little variation with
@@ -223,8 +215,8 @@ plot!(inputs[1, :], inputs[2, :]; seriestype = :scatter, zcolor = outputs[2, :],
 p = plot(
     p1,
     p2,
-    right_margin=3mm,
-    bottom_margin=3mm,
+    right_margin = 3mm,
+    bottom_margin = 3mm,
     size = (600, 300),
     layout = (1, 2),
     guidefontsize = 14,
@@ -258,9 +250,9 @@ p2 = contour(
 plot!(inputs[1, :], inputs[2, :]; seriestype = :scatter, zcolor = outputs[2, :], label = :false)
 p = plot(
     p1,
-    p2, 
-    right_margin=3mm,
-    bottom_margin=3mm,
+    p2,
+    right_margin = 3mm,
+    bottom_margin = 3mm,
     size = (600, 300),
     layout = (1, 2),
     guidefontsize = 14,
@@ -305,8 +297,8 @@ p = plot(
     p2,
     size = (600, 300),
     layout = (1, 2),
-    right_margin=3mm,
-    bottom_margin=3mm,
+    right_margin = 3mm,
+    bottom_margin = 3mm,
     guidefontsize = 14,
     tickfontsize = 12,
     legendfontsize = 12,
@@ -337,15 +329,7 @@ p2 = contour(
     ylabel = "Vertical Shift",
     title = "RF 1σ in Sinusoid Mean",
 )
-p = plot(
-    p1,
-    p2,
-    size = (600, 300),
-    layout = (1, 2),
-    guidefontsize = 14,
-    tickfontsize = 12,
-    legendfontsize = 12,
-)
+p = plot(p1, p2, size = (600, 300), layout = (1, 2), guidefontsize = 14, tickfontsize = 12, legendfontsize = 12)
 savefig(p, joinpath(data_save_directory, "sinusoid_RF_emulator_std_contours.png"))
 # The GP and RF uncertainty predictions are similar and show lower uncertainties around the region of interest
 # where we have more training points.
@@ -377,15 +361,7 @@ p2 = contour(
     ylabel = "Vertical Shift",
     title = "GP error in Sinusoid Mean",
 )
-p = plot(
-    p1,
-    p2,
-    size = (600, 300),
-    layout = (1, 2),
-    guidefontsize = 14,
-    tickfontsize = 12,
-    legendfontsize = 12,
-)
+p = plot(p1, p2, size = (600, 300), layout = (1, 2), guidefontsize = 14, tickfontsize = 12, legendfontsize = 12)
 savefig(p, joinpath(data_save_directory, "sinusoid_GP_errors_contours.png"))
 
 rf_diff_grid = abs.(rf_grid - g_true_grid)
@@ -411,15 +387,7 @@ p2 = contour(
     ylabel = "Vertical Shift",
     title = "RF error in Sinusoid Mean",
 )
-p = plot(
-    p1,
-    p2,
-    size = (600, 300),
-    layout = (1, 2),
[…]
