+++ examples/Sinusoid/sinusoid_setup.jl
@@ -82,7 +82,7 @@ function generate_obs(amplitude_true, vert_shift_true, Γ; rng = Random.GLOBAL_R
         guidefontsize = 14,
         tickfontsize = 12,
         legendfontsize = 12,
-)
+    )
 
     # We will observe properties of the signal that inform us about the amplitude and vertical 
     # position. These properties will be the range (the difference between the maximum and the minimum),
