+++ paper.bib
@@ -312,7 +312,7 @@ title = {PyVBMC: Efficient Bayesian inference in Python},
 journal = {Journal of Open Source Software}
 }
 
-@article{Gammal:2023
+@article{Gammal:2023,
 doi = {10.1088/1475-7516/2023/10/021},
 year = {2023},
 month = {oct},
