### 2024-04-23 69f9976
reduce trimmed fonts
