### 2024-04-23 332101c
JOSS address editor comments (#305)
* DOIs, resolved except RR2007, ScikitLearn, NUTS

* add Andre ORCID

* addressed e.g.s, reseearch list, and CES pipeline list

* replot sinusoid figures with larger fonts

* add trailing commas

* julia format

* trailing comma

* pdf update

* formatting bullets

* capitalization

* grammar...

* reduce trimmed fonts

* new pdf
