+++ paper.bib
@@ -1,6 +1,5 @@
 @article{julia,
   doi       = {10.1137/141000671},
-  url       = {https://doi.org/10.1137%2F141000671},
   year      = 2017,
   month     = {jan},
   publisher = {Society for Industrial {\&} Applied Mathematics ({SIAM})},
@@ -17,7 +16,8 @@ title = {Handbook of {Approximate} {Bayesian} {Computation}},
 isbn = {978-1-351-64346-7},
 publisher = {CRC Press},
 author = {Sisson, Scott A. and Fan, Yanan and Beaumont, Mark},
-year = {2018}
+year = {2018},
+doi = {10.1201/9781315117195}
 }
 
 @incollection{Nott:2018,
@@ -29,6 +29,7 @@ year = {2018}
   pages   = {211-241},
   year    = {2018},
   publisher = {CRC Press},
+  doi = {10.1201/9781315117195-8}
 }
 
 @article{Cleary:2021,
@@ -38,14 +39,12 @@ volume = {424},
 pages = {109716},
 year = {2021},
 issn = {0021-9991},
-doi = {https://doi.org/10.1016/j.jcp.2020.109716},
-url = {https://www.sciencedirect.com/science/article/pii/S0021999120304903},
+doi = {10.1016/j.jcp.2020.109716},
 author = {Emmet Cleary and Alfredo Garbuno-Inigo and Shiwei Lan and Tapio Schneider and Andrew M. Stuart},
 }
 
 @article{Dunbar:2022a,
 doi = {10.21105/joss.04869},
-url = {https://doi.org/10.21105/joss.04869},
 year = {2022},
 publisher = {The Open Journal},
 volume = {7},
@@ -72,7 +71,7 @@ year = {2022}
     school   = {Massachusetts Institute of Technology. Department of Electrical Engineering and Computer Science},
     author   = {Adeline Hillier},
     year     = {2022},
-    url = {https://hdl.handle.net/1721.1/145140}
+    doi = {1721.1/145140}
 }
 
 
@@ -82,7 +81,8 @@ year = {2022}
   volume={2},
   number={3},
   year={2006},
-  publisher={MIT press Cambridge, MA}
+  publisher={MIT press Cambridge, MA},
+  doi = {10.1142/S0129065704001899}
 }
 
 @article{Iglesias:2013,
@@ -93,7 +93,8 @@ year = {2022}
   number={4},
   pages={045001},
   year={2013},
-  publisher={IOP Publishing}
+  publisher={IOP Publishing},
+  doi={10.1088/0266-5611/29/4/045001}
 }
 
 @inproceedings{Rahimi:2007,
@@ -104,7 +105,7 @@ year = {2022}
   number={4},
   pages={5},
   year={2007},
-  organization={Citeseer}
+  url = {https://proceedings.neurips.cc/paper_files/paper/2007/file/013a006f03dbc5392effeb8f18fda755-Paper.pdf},
 }
 
 @inproceedings{Rahimi:2008,
@@ -113,7 +114,8 @@ year = {2022}
   booktitle={2008 46th Annual Allerton Conference on Communication, Control, and Computing},
   pages={555--561},
   year={2008},
-  organization={IEEE}
+  organization={IEEE},
+  doi = {10.1109/allerton.2008.4797607}
 }
 
 @article{Liu:2022,
@@ -138,21 +140,19 @@ pages = {424 -- 446},
 keywords = {algorithms, Bayesian inverse problems, Bayesian nonparametrics, Gaussian random field, MCMC},
 year = {2013},
 doi = {10.1214/13-STS421},
-URL = {https://doi.org/10.1214/13-STS421}
 }
 
 @article{Sherlock:2010,
[…]
