### 2024-04-23 #6372 @odunbar
@sappelhoff I've addressed your comments.

Please see https://github.com/CliMA/CalibrateEmulateSample.jl/pull/305 for the details of line-by-line changes etc.

The points I was unable to (fully) address
- three DOIs are in ML proceedings (NIPS/NEURIPS and JMLR) and do not provide DOIs. I could only find e.g. one with a preprint on arxiv (and the preprint has a DOI)
- The author would like to retain their name as `N. Efrat-Henrici`
