### 2024-04-23 #6372 @sappelhoff
Thanks a lot @odunbar, I'll have a look later today.

> PS Do I need to release a new version and get a new Zenodo link now or is this OK?

nope, this is fine as is, because the changes were all more in terms of the paper.
