### 2024-04-24 db7721b
add coeffs, hardcode some restart options in scalar RF
