### 2024-04-29 #6372 @arfon
@xuanxu – I wonder if we need some kind of 'accept ignoring errors' command here? The XML validation is overly-prescriptive here (i.e., wrong!)
