### 2024-04-29 #6372 @xuanxu
@arfon We can add that command yeah, but we have to take into account that this error comes from the Crossref XML schema validator so it means that if we ignore the errors we will try to deposit with Crossref a metadata file that they will consider invalid (I'm not sure if they will accept it anyway or reject the deposit).
