### 2024-05-03 #6372 @xuanxu
@arfon it should be `accept ignoring errors`
