### 2024-05-06 #6372 @danielskatz
It seems like this should be reported to Crossref, since the DOI does resolve.
