### 2024-05-06 #6372 @xuanxu
> Crossref failed to deposit the DOI :-(

It looks like it didn't fail, the Crossref API call response was OK:  https://github.com/openjournals/joss-papers/actions/runs/8973048619/job/24642364367#step:7:76
