### 2024-05-07 #6372 @arfon
> It seems like this should be reported to Crossref, since the DOI does resolve.

Crossref did OK. We knowingly submitted slightly bad XML (which it rejected). Waiting on an update from the author (@odunbar) to their BibTeX for this to be fixed.
