### 2024-05-07 #6372 @arfon
> How was our XML bad? If it was just that we have a DOI that Crossref doesn't like, that's what I think should be reported to Crossref, as the DOI is valid.

Right. It's the Crossref validation getting caught up on `1721.1/145140`
