+++ paper.bib
@@ -70,8 +70,7 @@ year = {2022}
     title    = {Supervised Calibration and Uncertainty Quantification of Subgrid Closure Parameters using Ensemble {Kalman} Inversion},
     school   = {Massachusetts Institute of Technology. Department of Electrical Engineering and Computer Science},
     author   = {Adeline Hillier},
-    year     = {2022},
-    doi = {1721.1/145140}
+    year     = {2022}
 }
 
 
