### 2024-05-07 #6372 @odunbar
Sorry - missed a lot of this discussion. I have merged https://github.com/CliMA/CalibrateEmulateSample.jl/pull/307
